StackScreen 구조 변경 및 버그 픽스
