+++ src/screens/Profile/index.tsx
@@ -1,17 +1,79 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
+import React, { useLayoutEffect, useState } from 'react';
+import { Alert, TouchableHighlight, View } from 'react-native';
+import { useDispatch, useSelector } from 'react-redux';
 
+import { Icon } from 'native-base';
+
+import { useNavigation } from '@react-navigation/native';
+
+import { removeRequesterToken } from '@/apis/BaseInstance';
+import { logout } from '@/apis/UserApi';
+import { Button } from '@/components';
+import { asyncStoragekey } from '@/constants/asyncStorage';
+import { StringStorage } from '@/helpers/functions/asyncStorage';
 import { RootState } from '@/store';
+import { clearAccessToken } from '@/store/userSlice';
+import { typo } from '@/styles';
 
 import LoggedProfile from './Logged';
-import UnloggedProfile from './UnLogged';
+import styles from './Profile.style';
+import UnLoggedProfile from './UnLogged';
 
 function Profile(): JSX.Element {
-  const isTokenExists = !!useSelector(
-    (state: RootState) => state.user.accessToken
+  const [show, setShow] = useState<boolean>(false);
+  const isTokenExists = useSelector(
+    (state: RootState) => !!state.user.accessToken
   );
 
-  return isTokenExists ? <LoggedProfile /> : <UnloggedProfile />;
+  const navigation = useNavigation();
+  const dispatch = useDispatch();
+
+  const logoutReq = async () => {
+    await logout();
+    await StringStorage.remove(asyncStoragekey.REFRESH_TOKEN);
+    dispatch(clearAccessToken());
+    removeRequesterToken();
+    Alert.alert('로그아웃 되었습니다.');
+  };
+
+  useLayoutEffect(() => {
+    navigation.setOptions({
+      title: '더보기',
+      headerTitleAlign: 'center',
+      // eslint-disable-next-line react/display-name
+      headerRight: () =>
+        isTokenExists ? (
+          <View style={{ position: 'relative' }}>
+            <TouchableHighlight onPress={() => setShow(!show)}>
+              <Icon name="menu" />
+            </TouchableHighlight>
+            {show ? (
+              <View style={styles.headerRightModal}>
+                <Button
+                  title="수정하기"
+                  textStyle={{
+                    ...typo.bigTitle,
+                  }}
+                  onPress={() => {
+                    setShow(false);
+                    navigation.navigate('ProfileModify');
+                  }}
+                />
+                <Button
+                  title="로그아웃하기"
+                  textStyle={{
+                    ...typo.bigTitle,
+                  }}
+                  onPress={logoutReq}
+                />
+              </View>
+            ) : null}
+          </View>
+        ) : null,
+    });
+  }, [isTokenExists, logoutReq, navigation, show]);
+
+  return isTokenExists ? <LoggedProfile /> : <UnLoggedProfile />;
 }
 
 export default Profile;
