+++ src/screens/StackScreens.tsx
@@ -112,132 +112,15 @@ function HomeStackScreen(): JSX.Element {
   );
 }
 
-// FIXME: @woohm402
-//   todo: 그때 말했던 폴더 구조 관련된 큰 체인지가 필요해 보입니다
-//         일단 어디로든 옮겨질 파일이라 생각하고 여러 군데 흩뿌려놓으면 나중에 찾기 힘드니까 여기다 다 몰아놓을게요
-//         현재 더보기창 디자인도 진행중인 관계로 정확하게 디자인하진 않겠습니당
-//   when: 폴더 구조 회의 완료되면
 const ProfileStack = createStackNavigator();
 function ProfileStackScreen(): JSX.Element {
-  const [show, setShow] = useState(false);
-  const logged = !!useSelector((state: RootState) => state.user.accessToken);
-  const navigation = useNavigation();
-  const dispatch = useDispatch();
-  const queryClient = useQueryClient();
-
-  const logoutReq = async () => {
-    await logout();
-    await StringStorage.remove(asyncStoragekey.REFRESH_TOKEN);
-    dispatch(clearAccessToken());
-    removeRequesterToken();
-    Alert.alert('로그아웃 되었습니다.');
-  };
-
-  const modifyUserProfileMutation = useMutation((values: IUserModify) => {
-    return modifyMyInfo(values);
-  });
-
-  const profileModifyFormik = useFormik<IUserModify>({
-    initialValues: {
-      nickname: '',
-      password: '',
-      trading_address: '',
-    },
-    validate: (values: IUserModify) => {
-      const errors = {};
-      // nickname
-      const nickname =
-        values.nickname &&
-        !isValidNickname(values.nickname) &&
-        '필수정보입니다.';
-      if (nickname) Object.assign(errors, { nickname });
-
-      return errors;
-    },
-    onSubmit: () => {
-      // 아 이 옵션 어떻게 안넣는방법 없나
-    },
-  });
-
   return (
     <ProfileStack.Navigator>
-      <ProfileStack.Screen
-        name={Profile.name}
-        component={Profile.component}
-        options={{
-          title: '더보기',
-          headerTitleAlign: 'center',
-          // eslint-disable-next-line react/display-name
-          headerRight: () =>
-            logged ? (
-              <View style={{ position: 'relative' }}>
-                <TouchableHighlight onPress={() => setShow(!show)}>
-                  <Icon name="menu" />
-                </TouchableHighlight>
-                {show ? (
-                  <View
-                    style={{
-                      position: 'absolute',
-                      top: 40,
-                      width: 169,
-                      right: 2,
-                      backgroundColor: 'white',
-                      height: 127,
-                      borderRadius: 10,
-                      borderWidth: 1,
-                      borderColor: palette.borderGray,
-                      paddingLeft: 33,
-                      justifyContent: 'space-evenly',
-                    }}
-                  >
-                    <Button
-                      title="수정하기"
-                      textStyle={{
-                        ...typo.bigTitle,
-                      }}
-                      onPress={() => navigation.navigate('ProfileModify')}
-                    />
-                    <Button
-                      title="로그아웃하기"
-                      textStyle={{
-                        ...typo.bigTitle,
-                      }}
-                      onPress={logoutReq}
-                    />
-                  </View>
-                ) : null}
-              </View>
-            ) : null,
-        }}
-      />
+      <ProfileStack.Screen name={Profile.name} component={Profile.component} />
       <ProfileStack.Screen
         name={ProfileModify.name}
-        options={{
-          headerTitleAlign: 'center',
-          // eslint-disable-next-line react/display-name
-          headerRight: () => (
-            <Button
-              title="완료"
-              onPress={async () => {
-                if (profileModifyFormik.errors) {
-                  Alert.alert('올바른 정보를 입력해 주세요.');
-                  return;
-                }
-
-                await modifyUserProfileMutation.mutateAsync(
-                  profileModifyFormik.values
-                );
-                await queryClient.invalidateQueries(USER_DETAIL);
-                navigation.goBack();
-              }}
-            />
-          ),
-        }}
-      >
-        {(props) => (
-          <ProfileModify.component {...props} formik={profileModifyFormik} />
-        )}
-      </ProfileStack.Screen>
+        component={ProfileModify.component}
+      />
     </ProfileStack.Navigator>
   );
 }
